First commit of Moodstocks Notes for Nokia.
If you package all files (except for the Visual Studio zipped folder) in a zipped folder and then change the extension to .wgz it should install on any Symbian phone.

Large portions of the code refer to Nokia's WRTkit which helps to keep the interface usable across different phone form factors. This kit is no longer supported but addins helping you use it can still be found on forum Nokia and IDE plugins exists for eclipse and Visual Studio 2008.

The large Visual Studio zipped folder contains the full project as a visual studio solution which should open in VS2008 if you have the WRTkit addin installed.